requirements.txt
flask
python-dotenv
cassio
langchain-community
langchain-openai
langchain-core
tavily-python

